BREAKING: Switch from `rustc_serialize` (deprecated) to `serde`
The Rust maintainers filed a bug against all packages using
`rustc_serialize` asking them to switch to `serde`, so let's take care
of that now.

This _technically_ includes a breaking API change, because we change the
type of the error wrapped in `credentials::ErrorKind::Json`. But this
should affect almost nobody.